InstructionList component
