+++ src/Components/Pages/Recipe.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import { useState, useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 
 import Typography from "@mui/material/Typography";
@@ -9,12 +9,7 @@ import AccordionDetails from "@mui/material/AccordionDetails";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import Button from "@mui/material/Button";
 import TextField from "@mui/material/TextField";
-import MenuItem from "@mui/material/MenuItem";
-import Select from "@mui/material/Select";
 import HighlightOffIcon from "@mui/icons-material/HighlightOff";
-import InputAdornment from "@mui/material/InputAdornment";
-import IconButton from "@mui/material/IconButton";
-import ClearIcon from "@mui/icons-material/Clear";
 
 import {
   createKey,
@@ -39,6 +34,7 @@ import {
   renderTagList,
   renderTagControls,
 } from "../Utils/RecipeParts";
+import InstructionList from "../Utils/InstructionList";
 
 function Recipe(props) {
   const {
@@ -83,8 +79,6 @@ function Recipe(props) {
   const [openCreateBasicFoodDialog, setOpenCreateBasicFoodDialog] =
     useState(false);
   const [createBasicFood, setCreateBasicFood] = useState({});
-
-  const [newStep, setNewStep] = useState("");
   const [newIngredientId, setNewIngredientId] = useState(null);
 
   useEffect(() => {
@@ -170,34 +164,6 @@ function Recipe(props) {
     updateIngredients((_ingredients) => {
       delete _ingredients[ingredientId];
       return _ingredients;
-    });
-  };
-
-  const moveStep = (oldIndex, newIndex) => {
-    updateInstructions((_instructions) => {
-      const step = _instructions[oldIndex];
-      _instructions.splice(oldIndex, 1);
-      _instructions.splice(newIndex, 0, step);
-      return _instructions;
-    });
-  };
-  const addStep = () => {
-    updateInstructions((_instructions) => {
-      return _instructions.concat(newStep);
-    });
-    setNewStep("");
-    document.getElementById("newStepInput").focus();
-  };
-  const updateStep = (index, step) => {
-    updateInstructions((_instructions) => {
-      _instructions.splice(index, 1, step);
-      return _instructions;
-    });
-  };
-  const getRemoveStep = (index) => () => {
-    updateInstructions((_instructions) => {
-      _instructions.splice(index, 1);
-      return _instructions;
     });
   };
 
@@ -513,134 +479,6 @@ function Recipe(props) {
     );
   };
 
-  const renderInstructions = () => {
-    const { instructions = [] } = recipeEntry;
-
-    return (
-      <Accordion key={"instructions"} sx={{ width: "100%" }}>
-        <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-          <Typography variant="h6">Instructions</Typography>
-        </AccordionSummary>
-        <AccordionDetails>
-          <Stack spacing={isEditing ? 2 : 1}>
-            {instructions
-              .map((instructionText, index) => {
-                return (
-                  <Stack
-                    key={index}
-                    direction="row"
-                    alignItems="center"
-                    spacing={1}
-                  >
-                    {isEditing ? (
-                      <>
-                        <Select
-                          value={index}
-                          sx={{ minWidth: "64px", height: "40px" }}
-                          onChange={(event) => {
-                            moveStep(index, event.target.value);
-                          }}
-                          onClose={() => {
-                            setTimeout(() => {
-                              document.activeElement.blur();
-                            }, 100);
-                          }}
-                        >
-                          {instructions.map((t, i) => (
-                            <MenuItem key={i} value={i} disabled={i === index}>
-                              {i + 1}
-                            </MenuItem>
-                          ))}
-                        </Select>
-                        <TextField
-                          placeholder="Edit step"
-                          value={instructionText}
-                          onChange={(event) => {
-                            updateStep(index, event.target.value);
-                          }}
-                          size="small"
-                          fullWidth={true}
-                          variant="outlined"
-                        />
-                        <HighlightOffIcon
-                          color="secondary"
-                          onClick={getRemoveStep(index)}
-                        />
-                      </>
-                    ) : (
-                      <>
-                        <Typography sx={{ fontWeight: 700, width: "17px" }}>
-                          {index + 1}.
-                        </Typography>
-                        &nbsp;
-                        <Typography>{instructionText}</Typography>
-                      </>
-                    )}
-                  </Stack>
-                );
-              })
-              .concat(
-                isEditing ? (
-                  <Stack key={"addStep"} direction="row" alignItems="center">
-                    <>
-                      <span
-                        onClick={() => {
-                          if (!newStep) {
-                            document.getElementById("newStepInput").focus();
-                          }
-                        }}
-                      >
-                        <Button
-                          color="secondary"
-                          variant="outlined"
-                          size="small"
-                          onClick={addStep}
-                          sx={{ minWidth: "64px", height: "40px" }}
-                          disabled={!newStep}
-                        >
-                          <Typography>Add</Typography>
-                        </Button>
-                      </span>
-                      &nbsp; &nbsp;
-                      <TextField
-                        id="newStepInput"
-                        placeholder="Enter new step"
-                        onChange={(event) => {
-                          setNewStep(event.target.value);
-                        }}
-                        size="small"
-                        fullWidth={true}
-                        variant="outlined"
-                        value={newStep}
-                        InputProps={{
-                          endAdornment: newStep && (
-                            <InputAdornment position="end">
-                              <IconButton
-                                sx={{ color: "alt.main" }}
-                                onClick={() => {
-                                  setNewStep("");
-                                  document
-                                    .getElementById("newStepInput")
-                                    .focus();
-                                }}
-                                edge="end"
-                              >
-                                <ClearIcon />
-                              </IconButton>
-                            </InputAdornment>
-                          ),
-                        }}
-                      />
-                    </>
-                  </Stack>
-                ) : null
-              )}
-          </Stack>
-        </AccordionDetails>
-      </Accordion>
-    );
-  };
-
   const renderNotes = () => {
     const { notes = "" } = recipeEntry;
 
@@ -689,7 +527,11 @@ function Recipe(props) {
         <Stack key="contents" spacing={2} sx={{ width: "95%", marginTop: 3 }}>
           {renderName()}
           {renderIngredients()}
-          {renderInstructions()}
+          <InstructionList
+            instructions={recipeEntry.instructions || []}
+            setInstructions={updateInstructions}
+            editable={isEditing}
+          />
           {renderNotes()}
           {renderTags()}
         </Stack>
